Move block producer config and more validator consts
